+++ judo-sdk/src/main/java/com/judopay/model/CollectionRequest.java
@@ -5,15 +5,15 @@ import com.judopay.api.Request;
 import java.math.BigDecimal;
 
 /**
- * A Collection allows for the funds reserved in a pre-auth transaction to be collected.
+ * A CollectionRequest allows for the funds reserved in a pre-auth transaction to be collected.
  */
 public class CollectionRequest extends Request {
 
     private final String receiptId;
     private final BigDecimal amount;
 
-    public CollectionRequest(String yourPaymentReference, String receiptId, BigDecimal amount) {
-        super(yourPaymentReference);
+    public CollectionRequest(String receiptId, BigDecimal amount) {
+        super(true);
         this.receiptId = receiptId;
         this.amount = amount;
     }