+++ CHANGELOG.md
@@ -1,11 +1,22 @@
 # judoNative Android SDK change log
 
+## [5.1.2](https://github.com/judopay/Judo-Android/tree/5.1.2) (2016-02-29)
+
+**Bugs fixed:**
+ - VoidRequest, CollectionRequest and RefundRequest no longer accept a 'yourPaymentReference' constructor argument as the request can be detected as a duplicate if the same payment reference was used from the pre-auth or payment request.
+
+**Changes:**
+- When performing a token payment, the expiry date of the token card is now shown along with the card number last 4 digits.
+- Request classes renamed: PaymentRequest, TokenRequest, RegisterCardRequest, VoidRequest, RefundRequest, CollectionRequest.
+- CV2 card location icon now shown on launch of payment screen, previously was only shown after card number input.
+
 ## [5.1.1](https://github.com/judopay/Judo-Android/tree/5.1.1) (2016-02-11)
 
 **Implemented enhancements:**
 - Added support for voiding pre-auth transactions to ```JudoApiService```
 
 **Bugs fixed:**
+- Visa Credit, Visa Debit and Visa Electron now show Visa logo when performing a token payment or pre-auth
 - Fixed an issue where additional API requests would fail after an initial transaction request was performed.
 - Resolved an issue where making a token payment/pre-auth would not allow a 4 digit CIDV to be entered for an Amex card.
 - Resolved an issue where an incorrect error message was given if the required Activity Extras was not passed to a ```JudoActivity```.